Rename RootNavigation.routes.ts
